+++ include/unity/scopes/internal/zmq_middleware/ZmqMiddleware.h
@@ -92,6 +92,7 @@ public:
     ThreadPool* oneway_pool();
     ThreadPool* twoway_pool();
     int64_t locate_timeout() const noexcept;
+    int64_t registry_timeout() const noexcept;
 
 private:
     ObjectProxy make_typed_proxy(std::string const& endpoint,
@@ -134,6 +135,7 @@ private:
 
     int64_t twoway_timeout_;                    // Default timeout for twoway invocations
     int64_t locate_timeout_;                    // Timeout for registry locate()
+    int64_t registry_timeout_;                  // Timeout for registry operations other than locate()
 
     std::string public_endpoint_dir_;
     std::string private_endpoint_dir_;