+++ scoperegistry/scoperegistry.cpp
@@ -27,22 +27,23 @@
 #include <unity/scopes/internal/RuntimeConfig.h>
 #include <unity/scopes/internal/RuntimeImpl.h>
 #include <unity/scopes/internal/ScopeConfig.h>
-#include <unity/scopes/internal/ScopeMetadataImpl.h>
 #include <unity/scopes/internal/ScopeImpl.h>
+#include <unity/scopes/internal/ScopeMetadataImpl.h>
 #include <unity/scopes/ScopeExceptions.h>
 #include <unity/UnityExceptions.h>
 #include <unity/util/ResourcePtr.h>
 
+#include <boost/algorithm/string.hpp>
+#include <boost/filesystem/operations.hpp>
+
 #include <algorithm>
 #include <cassert>
-#include <iostream>
-#include <sstream>
 #include <fstream>
+#include <iostream>
 #include <map>
 #include <set>
 
-#include <boost/filesystem/path.hpp>
-#include <boost/filesystem/operations.hpp>
+#include <sys/stat.h>  // TODO: will not be needed once the calls to mkdir are removed
 
 using namespace scoperegistry;
 using namespace std;
@@ -140,23 +141,28 @@ map<string, string> find_local_scopes(string const& scope_installdir, string con
     map<string, string> overrideable_scopes;    // Scopes that the OEM can override
 
     auto config_files = find_install_dir_configs(scope_installdir, ".ini", error);
-    for (auto&& path : config_files)
+    for (auto&& pair : config_files)
     {
+        if (boost::ends_with(pair.second, "-settings.ini"))
+        {
+            // Don't try to parse settings metadata file as scope config file.
+            continue;
+        }
         try
         {
-            ScopeConfig config(path.second);
+            ScopeConfig config(pair.second);
             if (config.overrideable())
             {
-                overrideable_scopes[path.first] = path.second;
+                overrideable_scopes[pair.first] = pair.second;
             }
             else
             {
-                fixed_scopes[path.first] = path.second;
+                fixed_scopes[pair.first] = pair.second;
             }
         }
         catch (unity::Exception const& e)
         {
-            error("ignoring scope \"" + path.first + "\": configuration error:\n" + e.what());
+            error("ignoring scope \"" + pair.first + "\": configuration error:\n" + e.what());
         }
     }
 
@@ -198,15 +204,20 @@ map<string, string> find_click_scopes(map<string, string> const& local_scopes, s
         try
         {
             auto click_paths = find_install_dir_configs(click_installdir, ".ini", error);
-            for (auto&& path : click_paths)
+            for (auto&& pair : click_paths)
             {
-                if (local_scopes.find(path.first) == local_scopes.end())
+                if (boost::ends_with(pair.second, "-settings.ini"))
                 {
-                    click_scopes[path.first] = path.second;
+                    // Don't try to parse settings metadata file as scope config file.
+                    continue;
+                }
+                if (local_scopes.find(pair.first) == local_scopes.end())
+                {
+                    click_scopes[pair.first] = pair.second;
                 }
                 else
                 {
-                    error("ignoring non-overrideable scope config \"" + path.second + "\" in click directory " + click_installdir);
+                    error("ignoring non-overrideable scope config \"" + pair.second + "\" in click directory " + click_installdir);
                 }
             }
         }
@@ -230,11 +241,19 @@ void add_local_scope(RegistryObject::SPtr const& registry,
                      string const& scoperunner_path,
                      string const& config_file,
                      bool click,
-                     int timeout_ms)
+                     int timeout_ms,
+                     string const& data_dir)
 {
     unique_ptr<ScopeMetadataImpl> mi(new ScopeMetadataImpl(mw.get()));
     string scope_config(scope.second);
     ScopeConfig sc(scope_config);
+
+    // TODO: HACK: We create the scope's cache dir until we get a fancier Apparmor API.
+    //       Once we have that, the scope will be able to find out what its writable
+    //       directory is and create it itself if necessary.
+    string confinement_type = sc.confinement_type();
+    ::mkdir((data_dir + "/" + confinement_type).c_str(), 0700);
+    ::mkdir((data_dir + "/" + confinement_type + "/" + scope.first).c_str(), 0700);
 
     filesystem::path scope_path(scope_config);
     filesystem::path scope_dir(scope_path.parent_path());
@@ -243,8 +262,12 @@ void add_local_scope(RegistryObject::SPtr const& registry,
     mi->set_settings_definitions(VariantArray());
     try
     {
-        auto schema = IniSettingsSchema::create(settings_schema_path.native());
-        mi->set_settings_definitions(schema->definitions());
+        // File is optional, so don't generate noise if it isn't there.
+        if (filesystem::exists(settings_schema_path))
+        {
+            auto schema = IniSettingsSchema::create(settings_schema_path.native());
+            mi->set_settings_definitions(schema->definitions());
+        }
     }
     catch (std::exception const& e)
     {
@@ -332,13 +355,14 @@ void add_local_scopes(RegistryObject::SPtr const& registry,
                       string const& scoperunner_path,
                       string const& config_file,
                       bool click,
-                      int timeout_ms)
+                      int timeout_ms,
+                      string const& data_dir)
 {
     for (auto&& pair : all_scopes)
     {
         try
         {
-            add_local_scope(registry, pair, mw, scoperunner_path, config_file, click, timeout_ms);
+            add_local_scope(registry, pair, mw, scoperunner_path, config_file, click, timeout_ms, data_dir);
         }
         catch (unity::Exception const& e)
         {
@@ -378,11 +402,28 @@ int main(int argc, char* argv[])
         SignalThreadWrapper signal_handler_wrapper;
 
         // And finally creating our runtime.
-        RuntimeConfig rt_config(config_file);
-        RuntimeImpl::UPtr runtime = RuntimeImpl::create(rt_config.registry_identity(), config_file);
+        // TODO: We need to make sure that the cache directory for each scope
+        //       is created if it does not exist. Once we have a fancier Apparmor
+        //       API, the dir should be created by the scope itself, so data_dir
+        //       should be removed again then.
+        string data_dir;
 
-        string identity = runtime->registry_identity();
-        string ss_reg_id = runtime->ss_registry_identity();
+        string identity;
+        string ss_reg_id;
+        RuntimeImpl::UPtr runtime;
+        {
+            RuntimeConfig rt_config(config_file);
+            runtime = RuntimeImpl::create(rt_config.registry_identity(), config_file);
+
+            identity = runtime->registry_identity();
+            ss_reg_id = runtime->ss_registry_identity();
+
+            // TODO: We need to make sure that the cache directory for each scope
+            //       is created if it does not exist. Once we have a fancier Apparmor
+            //       API, the dir should be created by the scope itself, so this
+            //       should be removed again then.
+            data_dir = rt_config.data_directory();
+        } // Release memory for config parser
 
         // Collect the registry config data.
 
@@ -441,8 +482,10 @@ int main(int argc, char* argv[])
             local_scopes[scope_id] = argv[i];                   // operator[] overwrites pre-existing entries
         }
 
-        add_local_scopes(registry, local_scopes, middleware, scoperunner_path, config_file, false, process_timeout);
-        add_local_scopes(registry, click_scopes, middleware, scoperunner_path, config_file, true, process_timeout);
+        add_local_scopes(registry, local_scopes, middleware, scoperunner_path,
+                         config_file, false, process_timeout, data_dir);
+        add_local_scopes(registry, click_scopes, middleware, scoperunner_path,
+                         config_file, true, process_timeout, data_dir);
         if (ss_reg_id.empty())
         {
             error("no remote registry configured, only local scopes will be available");
@@ -453,35 +496,37 @@ int main(int argc, char* argv[])
         }
 
         // Configure watches for scope install directories
-        ScopesWatcher local_scopes_watcher(registry,
-                                           [registry, &middleware, &scoperunner_path, &config_file, process_timeout]
-                                           (pair<string, string> const& scope)
+        auto local_watch_lambda = [registry, &middleware, &scoperunner_path, &config_file, process_timeout, data_dir]
+                                  (pair<string, string> const& scope)
         {
             try
             {
-                add_local_scope(registry, scope, middleware, scoperunner_path, config_file, false, process_timeout);
+                add_local_scope(registry, scope, middleware, scoperunner_path,
+                                config_file, false, process_timeout, data_dir);
             }
             catch (unity::Exception const& e)
             {
                 error("ignoring installed scope \"" + scope.first + "\": cannot create metadata: " + e.what());
             }
-        });
+        };
+        ScopesWatcher local_scopes_watcher(registry, local_watch_lambda);
         local_scopes_watcher.add_install_dir(scope_installdir);
         local_scopes_watcher.add_install_dir(oem_installdir);
 
-        ScopesWatcher click_scopes_watcher(registry,
-                                           [registry, &middleware, &scoperunner_path, &config_file, process_timeout]
-                                           (pair<string, string> const& scope)
+        auto click_watch_lambda = [registry, &middleware, &scoperunner_path, &config_file, process_timeout, data_dir]
+                                  (pair<string, string> const& scope)
         {
             try
             {
-                add_local_scope(registry, scope, middleware, scoperunner_path, config_file, true, process_timeout);
+                add_local_scope(registry, scope, middleware, scoperunner_path,
+                                config_file, true, process_timeout, data_dir);
             }
             catch (unity::Exception const& e)
             {
                 error("ignoring installed scope \"" + scope.first + "\": cannot create metadata: " + e.what());
             }
-        });
+        };
+        ScopesWatcher click_scopes_watcher(registry, click_watch_lambda);
         click_scopes_watcher.add_install_dir(click_installdir);
 
         // Let's add the registry's state receiver to the middleware so that scopes can inform