+++ src/scopes/internal/zmq_middleware/ZmqMiddleware.cpp
@@ -103,6 +103,7 @@ ZmqMiddleware::ZmqMiddleware(string const& server_name, RuntimeImpl* runtime, st
 
         twoway_timeout_ = config.twoway_timeout();
         locate_timeout_ = config.locate_timeout();
+        registry_timeout_ = config.registry_timeout();
         public_endpoint_dir_ = config.endpoint_dir();
         private_endpoint_dir_ = public_endpoint_dir_ + "/priv";
         registry_endpoint_dir_ = public_endpoint_dir_;
@@ -753,6 +754,11 @@ int64_t ZmqMiddleware::locate_timeout() const noexcept
     return locate_timeout_;
 }
 
+int64_t ZmqMiddleware::registry_timeout() const noexcept
+{
+    return registry_timeout_;
+}
+
 ObjectProxy ZmqMiddleware::make_typed_proxy(string const& endpoint,
                                             string const& identity,
                                             string const& category,
@@ -857,7 +863,7 @@ shared_ptr<ObjectAdapter> ZmqMiddleware::find_adapter(string const& name, string
         endpoint = "ipc://" + endpoint_dir + "/" + name;
     }
 
-    shared_ptr<ObjectAdapter> a(new ObjectAdapter(*this, name, endpoint, mode, pool_size, idle_timeout));
+    auto a = make_shared<ObjectAdapter>(*this, name, endpoint, mode, pool_size, idle_timeout);
     a->activate();
     am_[name] = a;
     return a;