+++ test/gtest/scopes/Invocation/TestScope.h
@@ -24,15 +24,8 @@
 class TestScope : public unity::scopes::ScopeBase
 {
 public:
-    virtual void start(std::string const&) override;
-
-    virtual void stop() override;
-
-    virtual void run() override;
-
     virtual unity::scopes::SearchQueryBase::UPtr search(unity::scopes::CannedQuery const&,
                                                         unity::scopes::SearchMetadata const&) override;
-
     virtual unity::scopes::PreviewQueryBase::UPtr preview(unity::scopes::Result const&,
                                                           unity::scopes::ActionMetadata const&) override;
 };