+++ test/gtest/scopes/Runtime/PusherScope.h
@@ -24,17 +24,10 @@
 class PusherScope : public unity::scopes::ScopeBase
 {
 public:
-    virtual void start(std::string const&) override;
-
-    virtual void stop() override;
-
-    virtual void run() override;
-
     virtual unity::scopes::SearchQueryBase::UPtr search(unity::scopes::CannedQuery const&,
                                                         unity::scopes::SearchMetadata const&) override;
-
-    virtual unity::scopes::PreviewQueryBase::UPtr preview(unity::scopes::Result const& result,
-                                                          unity::scopes::ActionMetadata const& metadata) override;
+    virtual unity::scopes::PreviewQueryBase::UPtr preview(unity::scopes::Result const&,
+                                                          unity::scopes::ActionMetadata const&) override;
 };
 
 #endif